own component for topic link, bigger link targets
